- Fixed bug #6359: Fatal error (and PHP warning) when adding source to an   RSS export.
git-svn-id: file:///home/patrick.allaert/svn-git/ezp-repo/ezpublish/stable/3.5@11224 a01eee8c-daba-0310-acae-fa49f3023285

+++ kernel/rss/edit_functions.php
@@ -44,7 +44,8 @@ function storeRSSExport( &$Module, &$http, $publish = false )
     $config =& eZINI::instance( 'site.ini' );
     $cacheDir = $config->variable( 'FileSettings', 'VarDir' ).'/'.$config->variable( 'FileSettings', 'CacheDir' );
     $cacheFile = $cacheDir . '/rss/' . md5( $http->postVariable( 'Access_URL' ) ) . '.xml';
-    unlink( $cacheFile );
+    if ( file_exists( $cacheFile ) )
+        unlink( $cacheFile );
 
     /* Create the new RSS feed */
     for ( $itemCount = 0; $itemCount < $http->postVariable( 'Item_Count' ); $itemCount++ )
